moved the parralax manager
